+++ src/dbt_charts/__init__.py
@@ -1,0 +1,3 @@
+"""dbt-charts package."""
+
+__version__ = "0.1.0"